Fix readme for redundant descriptions
