+++ docker/Makefile
@@ -1,9 +1,11 @@
 SHELL=/bin/bash
+PRIVATE_REGISTRY:=docker.jjgong.cloud
 BASE_IMAGE:=chloechiu/mol_base:pt112-cu116-dev
 
-PROJECT_NAME:=molcraft
+PROJECT_NAME:=bfn4sbdd2
 TEAM_NAME:=bfn
-PUBLIC_IMAGE:=${PROJECT_NAME}
+PUBLIC_IMAGE:=${PRIVATE_REGISTRY}/${TEAM_NAME}/${PROJECT_NAME}
+
 HTTPS_PROXY:=${https_proxy}
 HTTP_PROXY:=${http_proxy}
 _USER := $(shell whoami)
